pyproject.toml
[project]
name = "littlelemon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "djoser>=2.3.3",
    "mysql>=0.0.3",
]
